requirements.txt
litemapy
Pillow
numpy
requests
